Re implement the Mining Cautious Feature
Again the solution is actually from Doggy Talent Next, but that's still part of my code/solution so it is safe to reuse it here :)
This is an improvement from the old solution because this still allows the dog to follows the owner when he is mining, the dog instead of stopping, he just continue to walk but don't walk in front of owner.

Update SophisticatedWolf.java to exclude MoveCancelGoal because it is no longer needed.

Desc change and remvoe debug